requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
dj-rest-auth==2.2.5
django-allauth==0.50.0
djangorestframework-simplejwt==5.2.0
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
filelock==3.8.0
gunicorn==20.1.0
identify==2.5.5
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
django-admin-interface==0.19.1
django-colorfield==0.7.1